Experimentation, Reproducibility & Deployment Pipeline/requirement.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
pyyaml
streamlit
requests
joblib